+++ server/devApi/inboundController.js
@@ -1,19 +1,53 @@
 const Chatroom = require('./../db/controllers/chatroom.js');
 const App = require('./../db/controllers/app.js');
 
-// const SocketIo = require('socket.io');
-// const io = new SocketIo(server, {path: '/api/chat'});
+const fakeRoom = {
+  id: 'chasestarr/GitTalk',
+  apps: {
+    read: {
+      'http://localhost:8002': true,
+      'http://localhost:8003': true,
+      'chasebot%dot%herokuapp%dot%com': true
+    },
+    write: {
+      'abc': true,
+      '123': true
+    }
+  }
+}
 
 function inbound(req, res) {
   const payload = req.body;
-  const err = _ensureRequired(payload);
-  if (err) res.status(400).json(err);
 
-  _verifyRoomWriteAccess(payload.room, payload.apiKey, (err, chatroom) => {
-    if (err) res.status(400).json(err);
+  if (!payload.apiKey) {
+    res.status(400).json({ err: 'apiKey not found in request payload' });
+  }
 
+  if (!payload.method) {
+    res.status(400).json({ err: 'method not found in request payload' });
+  }
+
+  if (!payload.room) {
+    res.status(400).json({ err: 'room not found in request payload' });
+  }
+
+  Chatroom.findOne(payload.room, (err, chatroom) => {
+    if (err) res.status(400).json({ err: 'room not found' });
+
+    // will replace this code with real check
+    if (!fakeRoom.apps.write[payload.apiKey]) {
+      res.status(400).json({ err: `not authorized to write to ${payload.room}` });
+    }
+
+    // real lookup... commented out for now while chatroom schema is updated
+    // if (!chatroom.apps.write[payload.apiKey]) {
+    //   res.status(400).json({ err: `not authorized to write to ${payload.room}` });
+    // }
+
+    // find app name
     App.findOneByKey(payload.apiKey, (err, app) => {
-      if (err) res.status(400).json({ err: 'app not found' });
+      // if (err) res.status(400).json({ err: 'app not found' });
+      if (err) console.log(err);
 
       const message = {
         type: !!payload.action.image ? 'image': 'text',
@@ -21,47 +55,14 @@ function inbound(req, res) {
         image: payload.action.image,
         text: payload.action.text,
         userAvatarUrl: payload.action.avatar
-      };
+      }
 
-      const room = chatroom;
+      const room = chatroom[0];
       room.messages.push(message);
-
-      // io.sockets.emit('new bc message', msg);
-
-      Chatroom.update(room, () => {});
-      // room.save();
+      room.save();
     });
 
     res.status(201).end();
-  });
-}
-
-function _ensureRequired(payload) {
-  if (!payload.apiKey) {
-    return { err: 'apiKey field not found in inbound payload' };
-  }
-
-  if (!payload.method) {
-    return { err: 'method field not found in inbound payload' };
-  }
-
-  if (!payload.room) {
-    return { err: 'room field not found in inbound payload' };
-  }
-}
-
-function _verifyRoomWriteAccess(room, key, cb) {
-  Chatroom.findOne(room, (err, chatroom) => {
-    if (err) cb({ err: `${room} not found` }, null);
-    console.log('chatroom', chatroom[0]);
-    console.log('apps', chatroom[0].apps[0]);
-    console.log('key', key);
-
-    if (!chatroom[0].apps[0].write[key]) {
-      cb({ err: `not authorized to write to ${room}` }, null);
-    }
-
-    cb(null, chatroom[0]);
   });
 }
